refactor: 재고 도메인 요청 및 응답 dto 패키지 변경
