+++ hub/src/main/java/com/klp/hub/inventory/application/InventoryService.java
@@ -12,9 +12,9 @@ import com.klp.hub.inventory.domain.repository.dto.InventoryReplenish;
 import com.klp.hub.inventory.domain.repository.exception.UniqueConstraintException;
 import com.klp.hub.inventory.exception.InventoryErrorCode;
 import com.klp.hub.inventory.infrastructure.kafka.producer.InventoryEventProducer;
-import com.klp.hub.inventory.presentation.dto.InventoryDeductResponse;
-import com.klp.hub.inventory.presentation.dto.InventoryReplenishResponse;
-import com.klp.hub.inventory.presentation.dto.InventoryResponse;
+import com.klp.hub.inventory.presentation.dto.response.InventoryDeductResponse;
+import com.klp.hub.inventory.presentation.dto.response.InventoryReplenishResponse;
+import com.klp.hub.inventory.presentation.dto.response.InventoryResponse;
 import java.util.List;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
