+++ hub/src/main/java/com/klp/hub/inventory/presentation/dto/request/InventoryReplenishRequest.java
@@ -1,4 +1,4 @@
-package com.klp.hub.inventory.presentation.dto;
+package com.klp.hub.inventory.presentation.dto.request;
 
 import com.klp.hub.inventory.application.dto.InventoryReplenishCommand;
 import jakarta.validation.Valid;
