+++ hub/src/main/java/com/klp/hub/inventory/presentation/dto/response/InventoryReplenishResponse.java
@@ -1,4 +1,4 @@
-package com.klp.hub.inventory.presentation.dto;
+package com.klp.hub.inventory.presentation.dto.response;
 
 public record InventoryReplenishResponse(
     Status status
