+++ hub/src/test/java/com/klp/hub/inventory/application/InventoryFacadeConcurrencyTest.java
@@ -11,8 +11,8 @@ import com.klp.hub.inventory.domain.event.OrderCreatedEvent.OrderItemDto;
 import com.klp.hub.inventory.domain.repository.InventoryRepository;
 import com.klp.hub.inventory.infrastructure.repository.InventoryIdempotencyJpaRepository;
 import com.klp.hub.inventory.infrastructure.repository.InventoryJpaRepository;
-import com.klp.hub.inventory.presentation.dto.InventoryDeductResponse;
-import com.klp.hub.inventory.presentation.dto.InventoryDeductResponse.Status;
+import com.klp.hub.inventory.presentation.dto.response.InventoryDeductResponse;
+import com.klp.hub.inventory.presentation.dto.response.InventoryDeductResponse.Status;
 import java.util.List;
 import java.util.Set;
 import java.util.UUID;
