+++ hub/src/test/java/com/klp/hub/inventory/application/InventoryServiceTest.java
@@ -22,10 +22,10 @@ import com.klp.hub.inventory.domain.repository.InventoryRepository;
 import com.klp.hub.inventory.domain.repository.exception.UniqueConstraintException;
 import com.klp.hub.inventory.exception.InventoryErrorCode;
 import com.klp.hub.inventory.infrastructure.kafka.producer.InventoryEventProducer;
-import com.klp.hub.inventory.presentation.dto.InventoryDeductResponse;
-import com.klp.hub.inventory.presentation.dto.InventoryReplenishResponse;
-import com.klp.hub.inventory.presentation.dto.InventoryReplenishResponse.Status;
-import com.klp.hub.inventory.presentation.dto.InventoryResponse;
+import com.klp.hub.inventory.presentation.dto.response.InventoryDeductResponse;
+import com.klp.hub.inventory.presentation.dto.response.InventoryReplenishResponse;
+import com.klp.hub.inventory.presentation.dto.response.InventoryReplenishResponse.Status;
+import com.klp.hub.inventory.presentation.dto.response.InventoryResponse;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
