+++ hub/src/test/java/com/klp/hub/inventory/presentation/controller/InventoryControllerTest.java
@@ -13,9 +13,9 @@ import com.klp.hub.global.exception.GlobalExceptionHandler;
 import com.klp.hub.global.filter.AuthorizationFilter;
 import com.klp.hub.inventory.application.InventoryFacade;
 import com.klp.hub.inventory.application.InventoryService;
-import com.klp.hub.inventory.presentation.dto.InventoryReplenishRequest;
-import com.klp.hub.inventory.presentation.dto.InventoryReplenishResponse;
-import com.klp.hub.inventory.presentation.dto.InventoryResponse;
+import com.klp.hub.inventory.presentation.dto.request.InventoryReplenishRequest;
+import com.klp.hub.inventory.presentation.dto.response.InventoryReplenishResponse;
+import com.klp.hub.inventory.presentation.dto.response.InventoryResponse;
 import java.util.List;
 import java.util.UUID;
 import org.junit.jupiter.api.DisplayName;
